Add multi selection mode
In multi selection mode, directories can also be selected with 'Space'.
Selection resets, when changing directory or reopening the file dialog.

File dialog needs to be closed with q in multi selection mode. This is not
completely optimal.

Also adds a legend below the file dialog, explaining the non-obvious bindings,
if the default bindings are being used.

Use multi selection and file filter in the demo.